Split out the different dataset managers.
